refactor(backend): приложение meetings переименовано в votings
