+++ clocks/api/urls.py
@@ -3,5 +3,5 @@ from django.urls import include, path
 urlpatterns = [
     path("user/", include("users.urls")),
     path("room/", include("rooms.urls")),
-    path("meeting/", include("meetings.urls")),
+    path("meeting/", include("votings.urls")),
 ]
