+++ clocks/votings/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from meetings.models import Meeting
+from votings.models import Meeting
 
 
 @admin.register(Meeting)
