+++ clocks/votings/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class MeetingsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "meetings"
+    name = "votings"
