+++ clocks/votings/tests/test_meetings_api.py
@@ -10,8 +10,8 @@ def test_start_meeting_creates_and_notifies(api_client, room):
     url = reverse("start_meeting")
     payload = {"room": room.id, "task_name": "Разработка архитектуры"}
 
-    with patch("meetings.views.RoomMessageService") as MockRMS, patch(
-        "meetings.views.DjangoChannelMessageSender"
+    with patch("votings.views.RoomMessageService") as MockRMS, patch(
+        "votings.views.DjangoChannelMessageSender"
     ):
         MockRMS.return_value.notify_meeting_started = MagicMock()
 
@@ -55,7 +55,7 @@ def test_get_meeting_returns_info(api_client, meeting):
 @pytest.mark.django_db
 def test_end_meeting_only_active_allowed_and_calls_end_meeting(api_client, meeting):
     url = reverse("end_meeting", kwargs={"pk": meeting.id})
-    with patch("meetings.views.end_meeting") as mock_end:
+    with patch("votings.views.end_meeting") as mock_end:
         mock_end.return_value = None
         resp = api_client.put(url, data={}, format="json")
         assert resp.status_code == 200
@@ -73,10 +73,10 @@ def test_end_meeting_only_active_allowed_and_calls_end_meeting(api_client, meeti
 def test_restart_meeting_calls_services_and_resets_meeting(api_client, meeting):
     url = reverse("restart_meeting", kwargs={"pk": meeting.id})
 
-    with patch("meetings.views.DjangoChannelMessageSender"), patch(
-        "meetings.views.RoomMessageService"
-    ) as MockRMS, patch("meetings.views.RoomCacheService") as MockRoomCache, patch(
-        "meetings.views.RoomOnlineTracker"
+    with patch("votings.views.DjangoChannelMessageSender"), patch(
+        "votings.views.RoomMessageService"
+    ) as MockRMS, patch("votings.views.RoomCacheService") as MockRoomCache, patch(
+        "votings.views.RoomOnlineTracker"
     ) as MockTracker:
         MockRMS.return_value.notify_meeting_restart = MagicMock()
         MockRoomCache.return_value.clear_votes = MagicMock()
@@ -111,10 +111,10 @@ def test_update_meeting_task_requires_authorization_and_notifies(api_client, mee
     token = jwt_token
     api_client.credentials(HTTP_AUTHORIZATION=f"Bearer {token}")
 
-    with patch("meetings.views.JWTService") as MockJWT, patch(
-        "meetings.views.RoomCacheService"
-    ) as MockRoomCache, patch("meetings.views.UserSessionService") as MockUSS, patch(
-        "meetings.views.RoomMessageService"
+    with patch("votings.views.JWTService") as MockJWT, patch(
+        "votings.views.RoomCacheService"
+    ) as MockRoomCache, patch("votings.views.UserSessionService") as MockUSS, patch(
+        "votings.views.RoomMessageService"
     ) as MockRMS:
 
         MockJWT.return_value.decode.return_value = {
@@ -154,7 +154,7 @@ def test_meeting_results_computes_and_returns_votes_and_average(api_client, meet
         meeting_obj.average_score = 8.0
         meeting_obj.save()
 
-    with patch("meetings.views.meeting_results", side_effect=fake_meeting_results) as mock_results:
+    with patch("votings.views.meeting_results", side_effect=fake_meeting_results) as mock_results:
         resp = api_client.put(url, data={}, format="json")
         assert resp.status_code == 200
         body = resp.json()
